refactor(bin): move init utils to separate file
